Fix build failures for python 3.8
This fix tries to address the issue raised in 33543 where
tensorflow build on python 3.8 failed.

This fix fixed the issue as was suggested in 33543 and pip
builds finished successfully.

NOTE: tensorflow depends on h5py which does not have python 3.8
support yet, as such a release version of tensorflow for python 3.8
may have to wait for h5py first.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>