Fix CVE-2021-37652
Cherry-picks upstream commit 5ecec9c6f.

Prevent use after free.

A very old version of the code used `result` as a simple pointer to a resource.
Two years later, the pointer got changed to a `unique_ptr` but author forgot to
remove the call to `Unref`. Three years after that, we finally uncover the UAF.

PiperOrigin-RevId: 387924872
Change-Id: I70fb6f199164de49fac20c168132a07b84903f9b