[ARM] Disable castElements on arm device.
`castElements` call dnnl::cvt_<type1>_to_<type2> function, which is not defined by original onednn, but defined from a patch file.
In the patch file, we defined type `bfloat16_t` and `float16_t`, which also be defined in `arm_neon.h`.
So, we will face redefined errors when build DeepRec on arm devices.
Now, we temporarily disable `castElements` function, which means we can't get simd acceleration on type conversion, but it works fine.